Create indel_pyvolve_simulator.py
simulator working fully - need to change hard paths but that is it. this simulator takes in a tree file - and will run indelible, take the sequence length , site rates and indel pattern. then will run pyvolve with the indel sequence length and for three different omega/matrices and return three fas sequences.  these fas sequences are then merged based on the site rate from indelible, and then the indel pattern is applies. This simulation allows for a simulation with multi hits/indels with a custom matrix/ omega values.